Changed The way Frames are read from camera. Fixed Clip class to calculate motion fields from REF and TAR instead of REF and REF. Refactored method Calculate3DDisplacements to CalculateMotionField. Fixed PPMImage read/write methods. Fixed PixelTracker to account for neighbour displacements if pixel has no correspondent on upper layer on change of pyramid level. Fixed PointSets to account for floating point UV coordinates (used for subpixel interpolation of motion fields). Fixed Wireframe mode toggle button. Changed Error score calculation to a finer sampling of images. Created methods to apply motion fields to a point cloud.

+++ include/Vertex.h
@@ -9,14 +9,14 @@ private:
     Vec3Df          m_position;
     Vec3Df          m_normal;
     Color           m_color;
-    unsigned int    m_U;
-    unsigned int    m_V;
+    float           m_U;
+    float           m_V;
 
 public:
     // Constructors & destructors
     Vertex(void);
     Vertex(const Vec3Df& iPosition, const Vec3Df& iNormal, const Color& iColor);
-    Vertex(const Vec3Df& iPosition, const Vec3Df& iNormal, const Color& iColor, const unsigned int& iU, const unsigned int& iV);
+    Vertex(const Vec3Df& iPosition, const Vec3Df& iNormal, const Color& iColor, const float& iU, const float& iV);
     Vertex(const Vertex& iSource);
     ~Vertex(void);
 
@@ -30,8 +30,8 @@ public:
     const Color& GetColor(void) const;
     void SetColor(const Color& iColor);
 
-    void GetUVCoord(unsigned int& oU, unsigned int& oV) const;
-    void SetUVCoord(const unsigned int& iU, const unsigned int& iV);
+    void GetUVCoord(float& oU, float& oV) const;
+    void SetUVCoord(const float& iU, const float& iV);
 
     // Operators
     Vertex& operator=(const Vertex& iOther);
